moved ListTokens API view to separate file
